backend/app.py
from flask import Flask, request, jsonify
from google.cloud import vision, texttospeech
import google.auth
from flask_cors import CORS

app = Flask(__name__)
CORS(app)

# Authenticate using service account key
vision_client = vision.ImageAnnotatorClient()
# Gemini client setup (this step is simplified, as Gemini APIs may have specialized methods)

@app.route('/analyze-image', methods=['POST'])
def analyze_image():

    if 'image' not in request.files:
        return jsonify({'error': 'No image uploaded'}), 400

    # Save the uploaded image
    image_file = request.files['image']
    image_content = image_file.read()

    # Analyze the image with Google Vision
    image = vision.Image(content=image_content)
    response = vision_client.label_detection(image=image)

    if response.error.message:
        return jsonify({'error': f"Vision API Error: {response.error.message}"}), 500

    # Extract labels
    labels = [label.description for label in response.label_annotations]

    # Use Gemini for text generation
    # (Assume GeminiClient is a placeholder for actual Gemini API calls)
    gemini_output = {
        "prompt": f"Describe an image containing: {', '.join(labels)}",
        "result": "This is a placeholder for Gemini's text output."
    }

    return jsonify({
        'labels': labels,
        'gemini_description': gemini_output["result"]
    })

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=True)

